admin/yaab_corp/requirements.txt
anyio==4.1.0
apimarket==2.2.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.5
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==2.1
exceptiongroup==1.2.0
freetype-py==2.3.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2pdf==0.1
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
hyperframe==6.0.1
idna==3.4
lxml==4.9.3
mifiel==0.0.13
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==10.1.0
psycopg2==2.9.9
pycairo==1.25.1
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.17.1
pypng==0.20220715.0
python-bidi==0.4.2
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.3.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
webencodings==0.5.1
xhtml2pdf==0.2.13
